+++ src/main/java/net/geforcemods/securitycraft/compat/ftbteams/FTBTeamsCompat.java
@@ -1,10 +1,14 @@
 package net.geforcemods.securitycraft.compat.ftbteams;
 
+import java.util.ArrayList;
+import java.util.Collection;
 import java.util.UUID;
 
 import dev.ftb.mods.ftbteams.FTBTeamsAPI;
 import dev.ftb.mods.ftbteams.data.Team;
 import net.geforcemods.securitycraft.api.Owner;
+import net.geforcemods.securitycraft.util.TeamUtils.TeamRepresentation;
+import net.minecraft.server.level.ServerPlayer;
 
 public class FTBTeamsCompat {
 	private FTBTeamsCompat() {}
@@ -28,5 +32,17 @@ public class FTBTeamsCompat {
 		catch (IllegalArgumentException e) {}
 
 		return null;
+	}
+
+	public static Collection<ServerPlayer> getOnlinePlayersInTeam(Owner owner) {
+		try {
+			Team team = FTBTeamsAPI.getPlayerTeam(UUID.fromString(owner.getUUID()));
+
+			if (team != null)
+				return team.getOnlineMembers();
+		}
+		catch (IllegalArgumentException e) {}
+
+		return new ArrayList<>();
 	}
 }